git-svn-id: https://chibios.svn.sourceforge.net/svnroot/chibios/branches/utils_dev@3119 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ eclipse_debug/src/org/chibios/tools/eclipse/debug/utils/HexUtils.java
@@ -87,7 +87,7 @@ public class HexUtils {
   static public int parseInt(String s) {
     
     if (s.toLowerCase().startsWith("0x"))
-      return Integer.parseInt(s.substring(2), 16);
-    return Integer.parseInt(s);
+      return (int)Long.parseLong(s.substring(2), 16);
+    return (int)Long.parseLong(s);
   }
 }
